torrent shit refactor
