+++ source-api/src/commonMain/kotlin/eu/kanade/tachiyomi/torrentutils/TorrentUtils.kt
@@ -1,22 +1,17 @@
 package eu.kanade.tachiyomi.torrentutils
 
-import eu.kanade.tachiyomi.data.torrentServer.TorrentServerApi
+import eu.kanade.tachiyomi.torrentServer.TorrentServerApi
 import eu.kanade.tachiyomi.torrentutils.model.DeadTorrentException
 import eu.kanade.tachiyomi.torrentutils.model.TorrentFile
 import eu.kanade.tachiyomi.torrentutils.model.TorrentInfo
 import java.net.SocketTimeoutException
 
-/**
- * Used by extensions.
- */
-@Suppress("UNUSED")
 object TorrentUtils {
-    suspend fun getTorrentInfo(
+    fun getTorrentInfo(
         url: String,
         title: String,
     ): TorrentInfo {
-        @Suppress("SwallowedException")
-        try {
+        @Suppress("SwallowedException") try {
             val torrent = TorrentServerApi.addTorrent(url, title, "", "", false)
             return TorrentInfo(
                 torrent.title,
